Select story. Changed base decorator story render as per recommendations to fix components states resetting when using controls
